[build, cmake] port to solaris (#96)
Reviewed-on: https://git.eden-emu.dev/eden-emu/eden/pulls/96
Co-committed-by: lizzie <lizzie@eden-emu.dev>

+++ src/core/hle/service/caps/caps_manager.cpp
@@ -1,3 +1,6 @@
+// SPDX-FileCopyrightText: Copyright 2025 Eden Emulator Project
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 // SPDX-FileCopyrightText: Copyright 2023 yuzu Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
@@ -440,8 +443,8 @@ Result AlbumManager::SaveImage(ApplicationAlbumEntry& out_entry, std::span<const
     const auto screenshot_path =
         Common::FS::GetEdenPathString(Common::FS::EdenPath::ScreenshotsDir);
     const std::string formatted_date =
-        fmt::format("{:04}-{:02}-{:02}_{:02}-{:02}-{:02}-{:03}", date.year, date.month, date.day,
-                    date.hour, date.minute, date.second, 0);
+        fmt::format("{:04}-{:02}-{:02}_{:02}-{:02}-{:02}-{:03}", u16(date.year), u8(date.month), u8(date.day),
+                    u8(date.hour), u8(date.minute), u8(date.second), 0);
     const std::string file_path =
         fmt::format("{}/{:016x}_{}.png", screenshot_path, title_id, formatted_date);
